+++ src/stowd/stowd.py
@@ -6,7 +6,7 @@ import sys
 from importlib.metadata import version
 
 from .options.args import getargs
-from .options.config import get_config, get_config_settings, get_platform
+from .options.config import get_config
 from .options.settings import get_settings
 from .stow.print_results import print_results
 from .stow.stow import stow_from_args, stow_from_config
@@ -28,11 +28,9 @@ def main() -> None:
 
     # config file
     config = get_config(args.config_file, args.dotfiles_dir)
-    platform = get_platform(args.platform, config)
-    config_settings = get_config_settings(config, platform)
 
     # set setting from (command line arguments > config file > default)
-    settings = get_settings(args, config_settings)
+    settings = get_settings(args, config["settings"])
 
     # print that we're in Simulation mode
     if settings["simulate"]:
@@ -45,7 +43,7 @@ def main() -> None:
 
     # [un]stow[-root] from config file
     if sum(counter) == 0 or args.platform is not None:
-        stow_from_config(config, platform, counter, settings)
+        stow_from_config(config["home"], config["root"], counter, settings)
 
     # print results
     if not settings["quiet"]: